+++ contracts/interfaces/IBridge.sol
@@ -4,5 +4,5 @@ interface IBridge {
 
     // not sure if this should be done, exposing getter for _ChainID only
     // better than using call
-    function _chainID() external returns (uint);
+    function _chainID() external returns (uint8);
 }
